convert AsyncWorker
