+++ Samples/AsyncWorker/Sources/FakeNetwork/FakeNetworkManager.swift
