+++ Samples/AsyncWorker/Sources/FakeNetwork/Model.swift
