+++ Samples/AsyncWorker/Sources/MessageViewController.swift
